Refactor some controls
